+++ public/index.html
@@ -7,25 +7,18 @@
     <title>Big Trip</title>
 
     <link href="https://fonts.googleapis.com/css?family=Montserrat:400,500,600,700,800&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="./css/style.css">
   </head>
 
   <body class="page-body">
     <header class="page-header">
       <div class="page-body__container  page-header__container">
         <img class="page-header__logo" src="img/logo.png" width="42" height="42" alt="Trip logo">
-
         <div class="trip-main">
-          <!-- Маршрут и стоимость -->
-
           <div class="trip-main__trip-controls  trip-controls">
             <h2 class="visually-hidden">Switch trip view</h2>
-            <!-- Меню -->
-
             <h2 class="visually-hidden">Filter events</h2>
-            <!-- Фильтры -->
           </div>
-
           <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>
         </div>
       </div>
@@ -34,13 +27,9 @@
       <div class="page-body__container">
           <section class="trip-events">
             <h2 class="visually-hidden">Trip events</h2>
-
-            <!-- Сортировка -->
-
-            <!-- Контент -->
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
